v1.2.0
Redone the UI. Now the chat frontend + backend works fully and fine. Not as modular as I'd like, but decent enough for now to move on to extra options. Also not yet done creating all this stuff from within, but if you can get your hands a _very little if at all_ bit messy with .yaml .tpl and .ndjson files then its easy as and _basically_ done. Backend logic of multi messaging needs to be implemented but this works as a chatbot for now